Make Login() take username and password arguments
This is instead of requiring that the user set them in the struct before
calling Login().

I also made the username and password struct fields private. Let me know
if you think the username field should remain public. It may be useful
for UIs. That can always change when needed, though. Until then, it's
probably best to keep the API as simple as possible. I can't think of a
good reason for the password to be accessed in a context other than
Login(), so I'm confident that field can remain private.