Refactor page slice generation
It seems that storing it doesn't do us any good - we regenerate it every
time we need it anyway. This makes things more obvious and readable, and
makes the code more concise by avoiding the many three-deep struct
accesses that were necessary.

This commit also renames a few things for clarity and brevity.